Add. Returns diagnostic records in case of async executing.
From MSDN
```
The application should process diagnostic records in the repeating loop of the original function.
If SQLGetDiagField or SQLGetDiagRec is called when an asynchronous function is executing,
it will return the current list of diagnostic records.
Each time the original function call is repeated, it clears previous diagnostic records.
```